refactor: re-group fragments
